Full project code refactoring
