--- LibraryAppProject/DeeplinksHandlers/DeeplinksEnums.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-
-enum Tab: String {
-    case libraries = "libraries"
-    case events = "events"
-    case info = "info"
-}
