+++ LibraryAppProject/HandlersAndControllers/Deeplinks/DeeplinksEnums.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+// MARK: - Tabs Enum
+
+enum Tab: String {
+    case libraries = "libraries"
+    case events = "events"
+// Uncomment after adding 'info' tab at ContenView
+//    case info = "info"
+}
