+++ LibraryAppProject/Views/TabPagesViews/MultiSectionPage/SectionsViews/InfoSectionView.swift
@@ -1,0 +1,18 @@
+// Preparation of an information section for further development
+// Not called on project build
+
+import SwiftUI
+
+// MARK: - Information Section View
+
+struct InfoSectionView: View {
+    
+    // MARK: - Body
+    
+    var body: some View {
+        VStack {
+            Text("Контент первого подраздела")
+                .font(.title)
+        }
+    }
+}
